requirements.txt
sqllite3
requests
bs4
datetime
flask
flask-sqlalchemy
werkzeug
pyjwt